requirements.txt
django>=1.8,<1.9
django-debug-toolbar==1.4
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==3.0
django-taggit
python-dateutil==2.5.3
django-bootstrap-toolkit==2.15.0
django-bootstrap3==7.0.0
django-registration-redux==1.4
nltk==3.2
pbr==1.8.1
Pillow==3.1.0
simplejson==3.8.2
six==1.10.0
stevedore==1.10.0
textblob==0.11.1
textstat==0.2
virtualenv==13.1.2
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
virtualenvwrapper-win==1.2.1
xmltodict==0.10.1
